pyproject.toml
[tool.poetry]
name = "fastapi-yaml"
version = "0.3.0-rc"
description = "Handle YAML data in FastAPI"
authors = ["Mohamed Cherkaoui <chermed@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_yaml"}]
license = "MIT"
homepage = "https://github.com/chermed/fastapi-yaml"

[tool.poetry.dependencies]
python = ">= 3.11"
fastapi = ">= 0.100.0"
pyyaml = ">= 6.0"
python-multipart = ">=0.0.6,<0.1.0"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
pre-commit = "^3.3.3"
httpx = "^0.24.1"
pytest-cov = "^4.1.0"
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
fix = true
